requirements.txt
asgiref==3.8.1
Django~=5.2.0
django-redis==5.4.0
idna==3.10
JSON-log-formatter==1.1
psycogreen==1.0.2
psycopg[binary,pool]
pytz==2025.1
redis==6.2.0
requests==2.32.3
sqlparse==0.5.3
wheel==0.45.1
ccxt==4.4.88
eventlet==0.39.1
protobuf==5.29.3
grpcio==1.70.0
grpcio-tools==1.70.0
ntplib==0.4.0
python-dateutil==2.9.0.post0
raven==6.10.0
pydantic==2.11.4
httpx==0.28.1
celery==5.5.2
tenacity==9.1.2
djangorestframework==3.16.0
django-environ==0.12.0
django-celery-beat==2.8.1
gunicorn==23.0.0
uvicorn==0.34.3